Updated Enemy AI to shoot on Timer Code Clean Up(LOTS OF IT)

+++ DM2122_Framework/Application/Source/Player.h
@@ -22,9 +22,9 @@ public:
 
 	int getWeaponDMG();
 	void ShieldRegen(float Rate = 1.0f);
-	int healthIconVecX;
-	int shieldIconVecX;
-	int abilityIconVecX;
+	float healthIconVecX;
+	float shieldIconVecX;
+	float abilityIconVecX;
 	
 
 protected:
